+++ Library/Exception.cs
@@ -12,6 +12,17 @@ namespace Recurly
         /// </summary>
         public Error[] Errors { get; private set; }
 
+        /// <summary>
+        /// A transaction error (if there is one present)
+        /// </summary>
+        public TransactionError TransactionError { get; private set; }
+
+        internal RecurlyException(Errors errors)
+        {
+            Errors = errors.ValidationErrors;
+            TransactionError = errors.TransactionError;
+        }
+
         internal RecurlyException(Error[] errors)
         {
             Errors = errors;
@@ -24,6 +35,13 @@ namespace Recurly
         internal RecurlyException(string message, RecurlyException innerException)
             : base(message, innerException)
         { }
+
+        internal RecurlyException(string message, Errors errors)
+            : base(message)
+        {
+            Errors = errors.ValidationErrors;
+            TransactionError = errors.TransactionError;
+        }
 
         internal RecurlyException(string message, Error[] errors)
             : base(message)